requirements.txt
flask[async]==3.0.3
cryptography==3.4.8
httpx==0.27.0
numpy==1.24.0
pennylane==0.35.1
psutil==6.0.0
waitress==3.0.0
bleach==6.1.0
argon2-cffi==23.1.0
Flask-WTF==1.2.1
markdown2==2.5.0
WTForms==3.1.2
geonamescache==2.0.0
backoff==1.10.0
pathlib==1.0.1
llama-cpp-python==0.3.14
python-oqs==0.12.0

